Refactor: achievements 모듈을 users모듈에 병합
