+++ src/auth/auth.module.ts
@@ -7,9 +7,8 @@ import { AuthController } from './auth.controller';
 import { SessionSerializer } from './session.serializer';
 import { FtStrategy } from './strategies/ft.strategy';
 import { AuthService } from './auth.service';
-import { AchievementsService } from 'src/achievements/achievements.service';
-import { AchievementsRepository } from 'src/achievements/repositories/achievement.repository';
-import { UserAchievementsRepository } from 'src/achievements/repositories/user-achievement.repository';
+import { AchievementsRepository } from 'src/users/repositories/achievement.repository';
+import { UserAchievementsRepository } from 'src/users/repositories/user-achievement.repository';
 
 @Module({
   imports: [
@@ -26,7 +25,6 @@ import { UserAchievementsRepository } from 'src/achievements/repositories/user-a
     SessionSerializer,
     UsersService,
     AuthService,
-    AchievementsService,
   ],
 })
 export class AuthModule {}
