+++ src/channels/channels.module.ts
@@ -8,9 +8,8 @@ import { UsersService } from 'src/users/users.service';
 import { ChatsRepository } from './repositories/chats.repository';
 import { EventsModule } from 'src/events/events.module';
 import { UsersRepository } from 'src/users/repositories/users.repository';
-import { AchievementsService } from 'src/achievements/achievements.service';
-import { AchievementsRepository } from 'src/achievements/repositories/achievement.repository';
-import { UserAchievementsRepository } from 'src/achievements/repositories/user-achievement.repository';
+import { AchievementsRepository } from 'src/users/repositories/achievement.repository';
+import { UserAchievementsRepository } from 'src/users/repositories/user-achievement.repository';
 
 @Module({
   imports: [
@@ -25,6 +24,6 @@ import { UserAchievementsRepository } from 'src/achievements/repositories/user-a
     EventsModule,
   ],
   controllers: [ChannelsController],
-  providers: [ChannelsService, UsersService, AchievementsService],
+  providers: [ChannelsService, UsersService],
 })
 export class ChannelsModule {}
