+++ src/dms/dms.module.ts
@@ -1,8 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AchievementsService } from 'src/achievements/achievements.service';
-import { AchievementsRepository } from 'src/achievements/repositories/achievement.repository';
-import { UserAchievementsRepository } from 'src/achievements/repositories/user-achievement.repository';
+import { AchievementsRepository } from 'src/users/repositories/achievement.repository';
 import { EventsModule } from 'src/events/events.module';
 import { FriendshipsRepository } from 'src/friendships/friendships.repository';
 import { FriendshipsService } from 'src/friendships/friendships.service';
@@ -11,6 +9,7 @@ import { UsersService } from 'src/users/users.service';
 import { DmsController } from './dms.controller';
 import { DmsRepository } from './dms.repository';
 import { DmsService } from './dms.service';
+import { UserAchievementsRepository } from 'src/users/repositories/user-achievement.repository';
 
 @Module({
   imports: [
@@ -24,11 +23,6 @@ import { DmsService } from './dms.service';
     EventsModule,
   ],
   controllers: [DmsController],
-  providers: [
-    DmsService,
-    UsersService,
-    FriendshipsService,
-    AchievementsService,
-  ],
+  providers: [DmsService, UsersService, FriendshipsService],
 })
 export class DmsModule {}
