+++ src/friendships/friendships.service.ts
@@ -6,8 +6,7 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { AchievementsService } from 'src/achievements/achievements.service';
-import { AchievementName } from 'src/achievements/constants/achievement-name.enum';
+import { AchievementName } from 'src/users/constants/achievement-name.enum';
 import { User } from 'src/users/entities/user.entity';
 import { UsersService } from 'src/users/users.service';
 import { FriendshipRole } from './friendship-role.enum';
@@ -21,7 +20,6 @@ export class FriendshipsService {
     @InjectRepository(FriendshipsRepository)
     private readonly friendshipsRepository: FriendshipsRepository,
     private readonly usersService: UsersService,
-    private readonly achievementsService: AchievementsService,
   ) {}
 
   async createFriendship(
@@ -152,11 +150,11 @@ export class FriendshipsService {
     }
 
     if (friendship.status === FriendshipStatus.ACCEPTED) {
-      await this.achievementsService.createUserAchievement(
+      await this.usersService.createUserAchievement(
         requester,
         AchievementName.FIRST_FRIEND,
       );
-      await this.achievementsService.createUserAchievement(
+      await this.usersService.createUserAchievement(
         addressee,
         AchievementName.FIRST_FRIEND,
       );
@@ -297,7 +295,7 @@ export class FriendshipsService {
       addressee,
     );
 
-    await this.achievementsService.createUserAchievement(
+    await this.usersService.createUserAchievement(
       requester,
       AchievementName.FIRST_BLOCK,
     );
