+++ src/users/constants/achievement-name.enum.ts
@@ -1,5 +1,7 @@
 export enum AchievementName {
-  NEW_COMMER = 'NEW_COMMER',
+  FIRST_GAME = 'FIRST_GAME',
   FIRST_WIN = 'FIRST_WIN',
   FIRST_LOSE = 'FIRST_LOSE',
+  FIRST_FRIEND = 'FIRST_FRIEND',
+  FIRST_BLOCK = 'FIRST_BLOCK',
 }
